requirements.txt
telethon
redis
python-decouple